source-notion/requirements.txt
# Notion Source Connector Dependencies
# =====================================

# HTTP client with connection pooling
requests>=2.28.0,<3.0.0

# Configuration validation
pydantic>=2.0.0,<3.0.0

# Type hints support (for older Python versions)
typing-extensions>=4.0.0,<5.0.0

# Date/time handling
python-dateutil>=2.8.0,<3.0.0
